chore: rename lib/deskflow/option_types => lib/deskflow/OptionTypes
