+++ src/lib/deskflow/Screen.h
@@ -12,7 +12,7 @@
 #include "deskflow/IScreen.h"
 #include "deskflow/KeyTypes.h"
 #include "deskflow/MouseTypes.h"
-#include "deskflow/option_types.h"
+#include "deskflow/OptionTypes.h"
 
 class IClipboard;
 class IPlatformScreen;
