+++ src/lib/server/Server.cpp
@@ -17,11 +17,11 @@
 #include "deskflow/DropHelper.h"
 #include "deskflow/FileChunk.h"
 #include "deskflow/IPlatformScreen.h"
+#include "deskflow/OptionTypes.h"
 #include "deskflow/PacketStreamFilter.h"
 #include "deskflow/Screen.h"
 #include "deskflow/StreamChunker.h"
 #include "deskflow/XDeskflow.h"
-#include "deskflow/option_types.h"
 #include "deskflow/protocol_types.h"
 #include "mt/Thread.h"
 #include "net/TCPSocket.h"
